# demo/src/openapi/model/exerciseDto.ts
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { TypeEntity } from './typeEntity';


export interface ExerciseDto { 
    name?: string;
    type?: TypeEntity;
    gifUrl?: string;
}