# demo/src/openapi/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/api/v1/user/{userId}": {
      "put": {
        "tags": [
          "user-controller"
        ],
        "operationId": "updateUser",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/exercise/{exerciseId}": {
      "put": {
        "tags": [
          "exercise-controller"
        ],
        "operationId": "updateExercise",
        "parameters": [
          {
            "name": "exerciseId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExerciseDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "exercise-controller"
        ],
        "operationId": "deleteExercise",
        "parameters": [
          {
            "name": "exerciseId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/exercise": {
      "get": {
        "tags": [
          "exercise-controller"
        ],
        "operationId": "getExercises",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ExerciseEntity"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "exercise-controller"
        ],
        "operationId": "addExercise",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExerciseDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ExerciseEntity"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/user": {
      "get": {
        "tags": [
          "user-controller"
        ],
        "operationId": "getUser",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserEntity"
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "user-controller"
        ],
        "operationId": "deleteUser",
        "parameters": [
          {
            "name": "username",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/exercise/getExercisesByType": {
      "get": {
        "tags": [
          "exercise-controller"
        ],
        "operationId": "getExerciseByType",
        "parameters": [
          {
            "name": "typeDto",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/TypeDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ExerciseEntity"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "RegisterDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "height": {
            "type": "integer",
            "format": "int32"
          },
          "kg": {
            "type": "integer",
            "format": "int32"
          },
          "age": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "TypeDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          }
        }
      },
      "ExerciseDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "type": {
            "$ref": "#/components/schemas/TypeEntity"
          },
          "gifUrl": {
            "type": "string"
          }
        }
      },
      "ExerciseEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "gifUrl": {
            "type": "string"
          },
          "workout": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WorkoutEntity"
            }
          },
          "type": {
            "$ref": "#/components/schemas/TypeEntity"
          }
        }
      },
      "TypeEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "exercises": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExerciseEntity"
            }
          }
        }
      },
      "WorkoutEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "date_start": {
            "type": "string",
            "format": "date-time"
          },
          "date_finish": {
            "type": "string",
            "format": "date-time"
          },
          "duration": {
            "type": "integer",
            "format": "int32"
          },
          "exercises": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExerciseEntity"
            }
          }
        }
      },
...
      "RoleEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32",
            "writeOnly": true
          },
          "name": {
            "type": "string"
          }
        }
      },
      "UserEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "height": {
            "type": "integer",
            "format": "int32"
          },
          "kg": {
            "type": "integer",
            "format": "int32"
          },
          "age": {
            "type": "integer",
            "format": "int32"
          },
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/RoleEntity"
          },
          "workouts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WorkoutEntity"
            }
          }
        }
      }
    }
  }
}